azurephotos/requirements.txt
flask
azure-identity
azure-storage-blob
azure-data-tables
Pillow
